+++ app/src/main/java/com/geckour/q/service/DropboxMediaRetrieveService.kt
@@ -10,10 +10,10 @@ import android.content.Intent
 import android.content.IntentFilter
 import android.content.pm.PackageManager
 import android.graphics.Bitmap
-import android.media.MediaMetadataRetriever
 import android.webkit.MimeTypeMap
 import androidx.preference.PreferenceManager
 import com.dropbox.core.DbxRequestConfig
+import com.dropbox.core.RateLimitException
 import com.dropbox.core.v2.DbxClientV2
 import com.dropbox.core.v2.files.FileMetadata
 import com.dropbox.core.v2.files.FolderMetadata
@@ -21,18 +21,16 @@ import com.geckour.q.App
 import com.geckour.q.BuildConfig
 import com.geckour.q.R
 import com.geckour.q.data.db.DB
-import com.geckour.q.data.db.model.Album
-import com.geckour.q.data.db.model.Artist
-import com.geckour.q.data.db.model.Track
 import com.geckour.q.presentation.LauncherActivity
 import com.geckour.q.presentation.main.MainActivity
 import com.geckour.q.util.QNotificationChannel
-import com.geckour.q.util.UNKNOWN
 import com.geckour.q.util.dropboxToken
+import com.geckour.q.util.getExtension
 import com.geckour.q.util.getNotificationBuilder
-import com.geckour.q.util.storeArtwork
 import kotlinx.coroutines.runBlocking
 import timber.log.Timber
+import java.io.File
+import java.io.FileOutputStream
 
 class DropboxMediaRetrieveService : IntentService(NAME) {
 
@@ -140,62 +138,61 @@ class DropboxMediaRetrieveService : IntentService(NAME) {
         seed: Long,
         bitmap: Bitmap
     ) {
-        var fileAndFolders = client.files().listFolder(root)
-        while (true) {
-            if (fileAndFolders.hasMore.not()) break
+        try {
+            var fileAndFolders = client.files().listFolder(root)
+            while (true) {
+                if (fileAndFolders.hasMore.not()) break
 
-            fileAndFolders = client.files().listFolderContinue(fileAndFolders.cursor)
-        }
+                fileAndFolders = client.files().listFolderContinue(fileAndFolders.cursor)
+            }
 
-        fileAndFolders.entries.asSequence().forEach {
-            when (it) {
-                is FolderMetadata -> {
-                    retrieveAudioFilePaths(
-                        context,
-                        db,
-                        it.pathLower,
-                        client,
-                        seed,
-                        bitmap
-                    )
-                }
-                is FileMetadata -> {
-                    if (it.name.isAudioFilePath) {
-                        runCatching {
-                            val existingUrl = client.sharing()
-                                .listSharedLinksBuilder()
-                                .withPath(it.pathLower)
-                                .start()
-                                .links
-                                .maxByOrNull { it.expires?.time ?: 0 }
-                                ?.url
-                            val url = (existingUrl ?: client.sharing()
-                                .createSharedLinkWithSettings(it.pathLower)
-                                .url).replace("://www", "://dl")
-                            db.storeMediaInfo(context, it, url)
-                        }.onSuccess {
-                            filesCount++
-                        }.onFailure {
-                            Timber.e(it)
-                            return@forEach
-                        }
-
-                        sendBroadcast(MainActivity.createProgressIntent(filesCount to -1))
-                        startForeground(
-                            NOTIFICATION_ID_RETRIEVE,
-                            getNotification(filesCount)
+            fileAndFolders.entries.asSequence().forEach {
+                when (it) {
+                    is FolderMetadata -> {
+                        retrieveAudioFilePaths(
+                            context,
+                            db,
+                            it.pathLower,
+                            client,
+                            seed,
+                            bitmap
                         )
+                    }
+                    is FileMetadata -> {
+                        if (it.name.isAudioFilePath) {
+                            runCatching {
+                                storeMediaInfo(context, client, it)
+                            }.onSuccess {
+                                filesCount++
+                            }.onFailure { t ->
+                                if (t is RateLimitException) {
+                                    Thread.sleep(t.backoffMillis)
+                                    storeMediaInfo(context, client, it)
+                                } else {
+                                    Timber.e(t)
+                                    return@forEach
+                                }
+                            }
+
+                            sendBroadcast(MainActivity.createProgressIntent(filesCount to -1))
+                            startForeground(
+                                NOTIFICATION_ID_RETRIEVE,
+                                getNotification(filesCount)
+                            )
+                        }
                     }
                 }
             }
+        } catch (e: RateLimitException) {
+            Thread.sleep(e.backoffMillis)
+            retrieveAudioFilePaths(context, db, root, client, seed, bitmap)
         }
     }
 
     private val String.isAudioFilePath: Boolean
         get() = MimeTypeMap.getSingleton()
-            .getMimeTypeFromExtension(
-                this.replace(Regex("^.+\\.(.+?)$"), "$1")
-            )?.contains("audio") == true
+            .getMimeTypeFromExtension(this.getExtension())
+            ?.contains("audio") == true
 
     private fun getNotification(progress: Int): Notification =
         getNotificationBuilder(QNotificationChannel.NOTIFICATION_CHANNEL_ID_RETRIEVER)
@@ -227,107 +224,34 @@ class DropboxMediaRetrieveService : IntentService(NAME) {
             )
             .build()
 
-    private fun DB.storeMediaInfo(
+    private fun storeMediaInfo(
         context: Context,
-        dropboxMetadata: FileMetadata,
-        url: String
+        client: DbxClientV2,
+        dropboxMetadata: FileMetadata
     ): Long =
         runBlocking {
-            val metadataRetriever = MediaMetadataRetriever()
-                .apply { setDataSource(url, emptyMap()) }
+            val existingUrl = client.sharing()
+                .listSharedLinksBuilder()
+                .withPath(dropboxMetadata.pathLower)
+                .start()
+                .links
+                .maxByOrNull { it.expires?.time ?: 0 }
+                ?.url
+            val url = (existingUrl ?: client.sharing()
+                .createSharedLinkWithSettings(dropboxMetadata.pathLower)
+                .url).replace("://www", "://dl")
 
-            val trackTitle =
-                metadataRetriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_TITLE)
-                    ?: UNKNOWN
-            val artistTitle =
-                metadataRetriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_ARTIST)
-                    ?: UNKNOWN
-            val albumTitle =
-                metadataRetriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_ALBUM)
-                    ?: UNKNOWN
-            val duration =
-                metadataRetriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_DURATION)
-                    .toLong()
-            val pastSongDuration =
-                trackDao().getDurationWithTitles(trackTitle, albumTitle, artistTitle) ?: 0
-            val artworkUriString = metadataRetriever.embeddedPicture?.storeArtwork(context)
-            val artistId = artistDao().upsert(
-                Artist(
-                    0,
-                    artistTitle,
-                    artistTitle,
-                    0,
-                    duration,
-                    artworkUriString
-                ),
-                pastSongDuration
-            )
-            val albumArtistTitle =
-                metadataRetriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_ALBUMARTIST)
-            val albumArtistId =
-                if (albumArtistTitle != null) {
-                    artistDao().upsert(
-                        Artist(
-                            0,
-                            albumArtistTitle,
-                            albumArtistTitle,
-                            0,
-                            duration,
-                            artworkUriString
-                        ),
-                        pastSongDuration
-                    )
-                } else null
-            val albumId = albumDao().upsert(
-                Album(
-                    0,
-                    albumArtistId ?: artistId,
-                    albumTitle,
-                    albumTitle,
-                    artworkUriString,
-                    albumArtistId != null,
-                    0,
-                    duration
-                ),
-                pastSongDuration
-            )
-            val composerTitle =
-                metadataRetriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_COMPOSER)
-                    ?: UNKNOWN
-            val trackNum =
-                metadataRetriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_CD_TRACK_NUMBER)
-                    ?.split('/')
-                    ?.firstOrNull()
-                    ?.toInt()
-            val discNum =
-                metadataRetriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_DISC_NUMBER)
-                    ?.split('/')
-                    ?.firstOrNull()
-                    ?.toInt()
+            val dirName = "audio"
+            val fileName = "temp_audio.${dropboxMetadata.pathLower.getExtension()}"
+            val dir = File(context.cacheDir, dirName)
+            val file = File(dir, fileName)
 
-            metadataRetriever.release()
+            if (file.exists()) file.delete()
+            if (dir.exists().not()) dir.mkdir()
 
-            trackDao().upsertDropboxTrack(
-                Track(
-                    0,
-                    dropboxMetadata.serverModified.time,
-                    albumId,
-                    artistId,
-                    albumArtistId,
-                    -1,
-                    url,
-                    trackTitle,
-                    trackTitle,
-                    composerTitle,
-                    composerTitle,
-                    duration,
-                    trackNum,
-                    discNum,
-                    0,
-                ),
-                albumId,
-                artistId,
-                pastSongDuration
-            )
+            FileOutputStream(file).use {
+                client.files().download(dropboxMetadata.pathLower).download(it)
+                file.storeMediaInfo(context, url, null, dropboxMetadata.serverModified.time)
+            }
         }
 }
