--- app/src/main/java/com/geckour/q/service/MediaRetrieveServiceConst.kt
@@ -1,6 +1,0 @@
-package com.geckour.q.service
-
-internal const val NOTIFICATION_ID_RETRIEVE = 300
-
-internal const val ACTION_CANCEL = "com.geckour.q.service.retrieve.cancel"
-internal const val KEY_CLEAR = "key_clear"
