+++ app/src/main/java/com/geckour/q/util/String.kt
@@ -1,0 +1,3 @@
+package com.geckour.q.util
+
+fun String.getExtension(): String = replace(Regex("^.+\\.(.+?)$"), "$1")
